BLE: Move GATT public header into ble/gatt
Move private headers GattServerEvents.h, GapEvents,h and GapType.h in relevant part of the source tree.
