+++ connectivity/FEATURE_BLE/source/generic/GapImpl.cpp
@@ -25,8 +25,8 @@
 #include "ble/Gap.h"
 #include "ble/SecurityManager.h"
 #include "source/pal/PalGap.h"
-#include "ble/common/ble/GapEvents.h"
-#include "ble/common/ble/GapTypes.h"
+#include "source/pal/GapEvents.h"
+#include "source/pal/GapTypes.h"
 #include "source/pal/PalGenericAccessService.h"
 #include "source/pal/PalEventQueue.h"
 #include "source/pal/PalGap.h"
