+++ connectivity/FEATURE_BLE/source/generic/GapImpl.h
@@ -37,8 +37,8 @@
 #include "ble/gap/Events.h"
 
 #include "source/pal/PalGap.h"
-#include "ble/common/ble/GapEvents.h"
-#include "ble/common/ble/GapTypes.h"
+#include "source/pal/GapEvents.h"
+#include "source/pal/GapTypes.h"
 #include "source/pal/PalEventQueue.h"
 #include "source/pal/PalConnectionMonitor.h"
 #include "source/pal/PalEventQueue.h"
