+++ connectivity/FEATURE_BLE/source/pal/GapEvents.h
@@ -19,7 +19,7 @@
 #ifndef BLE_PAL_GAP_MESSAGE_H_
 #define BLE_PAL_GAP_MESSAGE_H_
 
-#include "ble/common/ble/GapTypes.h"
+#include "GapTypes.h"
 #include "ble/common/BLETypes.h"
 
 namespace ble {
