+++ connectivity/FEATURE_BLE/source/pal/GapTypes.h
