+++ connectivity/FEATURE_BLE/source/pal/PalGenericAccessService.h
@@ -19,7 +19,7 @@
 #ifndef BLE_PAL_GENERIC_ACCESS_SERVICE_H_
 #define BLE_PAL_GENERIC_ACCESS_SERVICE_H_
 
-#include "ble/common/ble/GapTypes.h"
+#include "GapTypes.h"
 #include "ble/common/BLETypes.h"
 #include "ble/common/blecommon.h"
 #include "ble/Gap.h"
